reformat code based on Hoang's comments
